[vpr][base] moove calculate_device_util to stats
